Change pcap_set_timeout's to_ms parameter's type from i32 to usize
As documented in libpcap's pcap_set_timeout manual page zero or negative
values are undefined. And timeout value in unsigned integer makes more
sense anyway.

+++ luomu-libpcap/src/lib.rs
@@ -275,7 +275,7 @@ impl PcapBuilder {
     /// capture handle when the handle is activated to to_ms, which is in units of
     /// milliseconds.
     pub fn set_timeout(self, to_ms: Duration) -> Result<PcapBuilder> {
-        let timeout = i32::try_from(to_ms.as_millis()).unwrap_or(i32::MAX);
+        let timeout = usize::try_from(to_ms.as_millis()).unwrap_or(usize::MAX);
         pcap_set_timeout(&self.pcap_t, timeout)?;
         Ok(self)
     }
